Mention some of our sponsors.
We need to be able to acknowledge contributions from everyone somehow, whether contributing code directly, providing funding (or employment!) to developers/designers/doc writers, or other resources (e.g. hosting).

This will probably belong on our site (which doesn't exist yet). For the time being, I'm adding it as a PR here for the sake of being more transparent.