Always on autocomplete
In order to do this, we have to set completeSingle to false. Otherwise users
 will "overtype" the completions with their own text.

For example, let's say a user is typing "import" in Python:

Type `i` (list of completions)

Type `m` ("import" matches, gets filled out)

Type p and now the text reads:

```
importp
```

Switching to this mode makes us behave more like Visual Studio and Atom. The user
must confirm the match in order to accept it. For previous Jupyter users, this
may cause a rift because muscle memory tells them to use tab. This _is_ a different environment
so maybe that's alright.

My personal preference is to have this mode on by default.