Don't do autocomplete when the change is a newline
When a newline is inserted before existing code like in #409, the change event's
text field is exactly `["", ""]`. Here that gets filtered out.

While this "solves" the problem for new lines, inserting a space before an existing
also has this problem (of showing IPython globals). We'll need to visit this again
and see if there's a way to simplify for the "natural" thing to do for automatic
hinting like we're doing here.