fix dist lifecycle
When packaging the last few releases I noticed that the dist folder wasn't getting
created and I ended up having to run the dist manually. Turns out it was because of
the `set` used here, as it literally `set` NODE_ENV to the long string and didn't
actually run electron-packager.